+++ src/main/java/de/muspellheim/commons/fx/concurrent/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Muspellheim Commons FX
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+/**
+ * Contains dispatch queues for application and background.
+ */
+package de.muspellheim.commons.fx.concurrent;